refactor: create a hook useBreadCrumbs
