+++ src/components/DisplayProductInfo/DisplayProductInfo.tsx
@@ -5,10 +5,8 @@ import { getProductByKey } from '../../api/products';
 import { Card } from 'primereact/card';
 import { FIRST_INDEX } from '../../constants/common';
 import { covertPrice } from '../../utils/product';
-import { useLocation, useNavigate } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import { BreadCrumb } from 'primereact/breadcrumb';
-import { MenuItem } from 'primereact/menuitem';
-import { PAGES } from '../../constants/pages';
 import styles from './DisplayProductInfo.module.scss';
 import { ToggleButton, ToggleButtonChangeEvent } from 'primereact/togglebutton';
 import { useIsItemInCart } from '../../hooks/useItemCart';
@@ -29,9 +27,9 @@ import {
   asyncUpdateCartProductId,
   cartUserDraft,
 } from '../../api/cart';
+import { useBreadCrumbs } from '../../hooks/useBreadCrumbs';
 
 export function DisplayProductInfo(keyProduct: string): JSX.Element {
-  const location = useLocation();
   const [images, setImages] = useState<ImageSDK[]>();
   const [nameProduct, setNameProduct] = useState<string>();
   const [descriptionProduct, setDescriptionProduct] = useState<string>();
@@ -112,27 +110,15 @@ export function DisplayProductInfo(keyProduct: string): JSX.Element {
     return <img src={item.url} alt={item.label} style={{ width: '50%' }} />;
   };
 
-  const items: MenuItem[] = [];
-  const home: MenuItem = { icon: 'pi pi-home', url: '/' };
-
-  location.pathname.split('/').forEach(path => {
-    if (path === PAGES.catalog.key) {
-      items.push({ label: `${path}`, url: `/${path}` });
-    } else if (
-      path.length &&
-      (path === PAGES.accessories.key ||
-        path === PAGES.textiles.key ||
-        path === PAGES.cosmetics.key)
-    ) {
-      items.push({ label: `${path}`, url: `../${path}` });
-    } else if (path.length) {
-      items.push({ label: `${path}` });
-    }
-  });
+  const { itemsBreadCrumbs, home } = useBreadCrumbs();
 
   return (
     <>
-      <BreadCrumb model={items} home={home} className={styles.breadcrumb} />
+      <BreadCrumb
+        model={itemsBreadCrumbs}
+        home={home}
+        className={styles.breadcrumb}
+      />
       <div className={styles.wrapper}>
         <Galleria
           value={images}
