+++ src/pages/CatalogPage/Catalog.tsx
@@ -2,22 +2,20 @@ import { ProductProjection } from '@commercetools/platform-sdk';
 import { BreadCrumb } from 'primereact/breadcrumb';
 import { Button } from 'primereact/button';
 import { InputText } from 'primereact/inputtext';
-import { MenuItem } from 'primereact/menuitem';
 import { Slider, SliderChangeEvent } from 'primereact/slider';
 import { ToggleButton, ToggleButtonChangeEvent } from 'primereact/togglebutton';
 import { useEffect, useState } from 'react';
-import { useLocation, useSearchParams } from 'react-router-dom';
+import { useSearchParams } from 'react-router-dom';
 import { getProducts } from '../../api/products';
 import { ProductItem } from '../../components/Product';
 import { PRODUCTS_IN_PAGE } from '../../constants/common';
-import { PAGES } from '../../constants/pages';
+import { useBreadCrumbs } from '../../hooks/useBreadCrumbs';
 import { FilterParams } from '../../types/types';
 import { getPageCount, getPagesArray } from '../../utils/product';
 import styles from './Catalog.module.scss';
 
 export const Catalog = ({ ...options }): JSX.Element => {
   const idCategory = options.options.id;
-  const location = useLocation();
 
   const [searchParams, setSearchParams] = useSearchParams();
   const currentLocation = searchParams.get('page') || '1';
@@ -95,19 +93,7 @@ export const Catalog = ({ ...options }): JSX.Element => {
     setResetFilters(true);
   };
 
-  const items: MenuItem[] = [];
-  const home: MenuItem = { icon: 'pi pi-home', url: '/' };
-  const locationPage = location.search?.split('=')[1];
-
-  location.pathname.split('/').forEach(path => {
-    if (path === PAGES.catalog.key) {
-      items.push({ label: `${path}`, url: `/${path}` });
-    }
-    if (path.length && path !== PAGES.catalog.key) {
-      items.push({ label: `${path}`, url: `${path}` });
-    }
-  });
-  if (locationPage) items.push({ label: `page ${locationPage}` });
+  const { itemsBreadCrumbs, home } = useBreadCrumbs();
 
   return (
     <div className={styles.content_main}>
@@ -132,7 +118,6 @@ export const Catalog = ({ ...options }): JSX.Element => {
         </span>
         <div className="card flex justify-content-start">
           <div className="w-14rem">
-            {/* <span>Price</span> */}
             <div className="input-container">
               <p style={{ margin: '0px' }}>Price From</p>
               <InputText
@@ -213,7 +198,11 @@ export const Catalog = ({ ...options }): JSX.Element => {
           className="w-8rem"
         />
         <div className={styles.main}>
-          <BreadCrumb model={items} home={home} className={styles.breadcrumb} />
+          <BreadCrumb
+            model={itemsBreadCrumbs}
+            home={home}
+            className={styles.breadcrumb}
+          />
           <div className={styles.content}>
             {products?.map(data => <ProductItem {...data} key={data.id} />)}
           </div>
